api/requirements-api.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyarrow>=12.0.0  # For parquet support

# Data Analysis & Visualization
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
kaleido>=0.2.1  # For plotly static image export

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0  # PostgreSQL async driver
alembic>=1.12.0  # Database migrations
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Utilities
python-dateutil>=2.8.2

